src/pyproject.toml
[tool.poetry]
name = "moodtracker"
version = "0.1.0"
description = ""
authors = ["Cameron Hobbs <camerontech@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework-simplejwt = "^5.2.2"
Django = "^4.1.3"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
gunicorn = {extras = ["gevent"], version = "^20.1.0"}
twilio = "^7.15.4"
redis = "^4.3.5"
celery = {extras = ["sqs"], version = "^5.2.7"}
mysqlclient = "^2.1.1"
drf-spectacular = "^0.25.1"
pymemcache = "^4.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-django = "^4.5.2"
factory-boy = "^3.2.1"
pytest-factoryboy = "^2.5.0"
requests-mock = "^1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
